iiTraffic,iiUsageUnit - Had to change quota storage to type 'long long'. Integer value max hit in some scenarios. iiVolumeUsageProvider - Constants for cache expiry and conversion. Don't store iiFeed in cache if it's nil. It will throw an error if you try. iiViewController - Add more labels to make it a more useful demo

+++ VolumeUsageAPI/iiTraffic.h
@@ -13,7 +13,8 @@
 @interface iiTraffic : NSObject
 
 @property (nonatomic, assign) TrafficType trafficType;
-@property (nonatomic, assign) NSInteger used;
+// Since this is in bytes we can exceed the max integer range
+@property (nonatomic, assign) long long used;
 @property (nonatomic, assign) NSInteger quota;
 @property (nonatomic, assign) BOOL isShaped;
 @property (nonatomic, strong) NSString *shapedSpeed;
